clean code, move to components
