+++ src/components/Welcome/WelcomeContainer.tsx
@@ -1,8 +1,8 @@
 import { useNavigate, useParams } from 'react-router-dom';
 import { Button } from '@codegouvfr/react-dsfr/Button';
-import { useRemote } from '../../components/orchestrator/useRemote';
+import { useRemote } from '../orchestrator/useRemote';
 import { useContext } from 'react';
-import { loadSourceDataContext } from '../../components/loadSourceData/LoadSourceDataContext';
+import { loadSourceDataContext } from '../loadSourceData/LoadSourceDataContext';
 import { Skeleton } from '@mui/material';
 import { RespondantsList } from './RespondantsList';
 import { WelcomeQuestions } from './WelcomeQuestions'
@@ -27,14 +27,11 @@ export function WelcomeContainer() {
         <div className="fr-grid-row">
             <div className="fr-col-12">
                 <h1>Bienvenue sur l'{welcome.Enq_LibelleEnquete}</ h1>
-            </div>
-            <div className="fr-col-12">
                 <p className="fr-text--lead">{welcome.Enq_ObjectifsCourts}</ p>
-            </div>
-            <RespondantsList respondants={welcome.whoAnswers} />
-            <div className="fr-col-12 fr-mt-2w">
+                <RespondantsList respondants={welcome.whoAnswers} />
                 {/* To do: rediriger vers la première question */}
                 <Button
+                    className="fr-col-12 fr-mt-2w"
                     size="large"
                     linkProps={{
                         href: '/',
@@ -42,11 +39,7 @@ export function WelcomeContainer() {
                 >
                     Commencer
                 </Button>
-            </div>
-            <div className="fr-col-12 fr-mt-5w">
-                <h3>En savoir plus sur l'{welcome.Enq_LibelleEnquete}</ h3>
-            </div>
-            <div className="fr-col-12">
+                <h3 className="fr-mt-5w">En savoir plus sur l'{welcome.Enq_LibelleEnquete}</ h3>
                 <WelcomeQuestions welcome={metadata.Welcome} />
             </div>
         </div >
