+++ src/components/Welcome/WelcomeType.ts
