+++ src/components/Welcome/index.ts
@@ -1,0 +1,1 @@
+export { WelcomeContainer } from './WelcomeContainer';
