+++ src/pages/welcome/Welcome.tsx
@@ -2,7 +2,7 @@ import { useParams } from 'react-router-dom';
 import { Layout } from '../../components/layout';
 import { useDocumentTitle } from '../../useDocumentTitle';
 import { LoadFromApi } from '../../components/loadSourceData/LoadFromApi';
-import { WelcomeContainer } from "./WelcomeContainer"
+import { WelcomeContainer } from "../../components/Welcome"
 
 
 export function Welcome() {
